+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ New features
 * Automatically set PD in $SIZES for NONMEM models when needed
 * Automatically add a dummy DV column in NONMEM if needed. This will make it possible to do simulations in NONMEM without having a DV in the dataset.
 * Add :code:`modeling.export_model_files`
+* Add :code:`modeling.map_eta_parameters` for creation of mappings between connected individual parameters, etas or omegas.
 
 Bugfixes
 ========